Refactor FieldElement multiplication and squaring
This replaces most of the stack variables with either array accesses or
function calls, which the JVM should optimize at runtime. The resulting
code is significantly clearer, fits in a single window, and is easier to
understand.

this.t is also cached locally, to reduce the number of getfield opcodes.